choropleth cleanup
