+++ server/__init__.py
@@ -76,11 +76,10 @@ def create_app():
     app.register_blueprint(ranking.bp)
     app.register_blueprint(redirects.bp)
     app.register_blueprint(tools.bp)
-    from routes.api import (browser as browser_api, chart, choropleth, place as
+    from routes.api import (browser as browser_api, choropleth, place as
                             place_api, landing_page, ranking as ranking_api,
                             stats, translator)
     app.register_blueprint(browser_api.bp)
-    app.register_blueprint(chart.bp)
     app.register_blueprint(choropleth.bp)
     app.register_blueprint(factcheck.bp)
     app.register_blueprint(place_api.bp)
