--- static/js/tools/choropleth/choropleth_template.tsx
@@ -1,193 +1,0 @@
-/**
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import React, { Component } from "react";
-import ReactDOM from "react-dom";
-import { Menu } from "../statvar_menu/menu";
-import { ChoroplethMap } from "./choropleth";
-import { NoopStatsVarFilter, TimelineStatsVarFilter } from "../commons";
-import axios from "axios";
-import statsVarPathMap from "../../../data/statsvar_path.json";
-
-const DEFAULT_STAT_VAR = "Count_Person_Employed";
-
-/**
- * Generates choropleth map from API on pageload.
- */
-window.onload = () => {
-  // Generate page.
-  ReactDOM.render(
-    React.createElement(MainPane),
-    document.getElementById("main-pane")
-  );
-};
-
-class MainPane extends Component {
-  constructor(props: Record<string, unknown>) {
-    super(props);
-
-    // Redirect to basic url if none provided.
-    const urlParams = new URLSearchParams(window.location.search);
-    const statVarDcid = urlParams.has("statVar")
-      ? urlParams.get("statVar")
-      : DEFAULT_STAT_VAR;
-    const statVarNodePath = statsVarPathMap[statVarDcid].map((x: number) =>
-      x.toString()
-    );
-    if (window.location.search === "") {
-      window.location.search = `statVar=${statVarDcid}&pc=t&geoDcid=country/USA`;
-    } else if (!urlParams.has("geoDcid")) {
-      urlParams.set("geoDcid", "country/USA");
-      window.location.search = urlParams.toString();
-    }
-
-    // Bind functions as needed.
-    this._handleStatVarSelection = this._handleStatVarSelection.bind(this);
-
-    // Get default values for optional fields.
-    let isPerCapita = false;
-    if (urlParams.has("pc")) {
-      isPerCapita = ["true", "t", "1"].includes(
-        urlParams.get("pc").toLowerCase()
-      );
-    }
-
-    // Get all statistical variable available for the current subgeo.
-    axios
-      .get("/api/choropleth/child/statvars?dcid=" + urlParams.get("geoDcid"))
-      .then((resp) => {
-        if (resp.status === 200) {
-          const statVars: Set<string> = new Set();
-          resp.data.forEach((item) => statVars.add(item));
-          this.setState({
-            statsVarFilter: new TimelineStatsVarFilter(statVars),
-          });
-        }
-      });
-
-    const statsVarNodes = {};
-    statsVarNodes[statVarDcid] = { paths: [statVarNodePath] };
-    // Initialize state.
-    this.state = {
-      // References used for both choropleth and stats var sidemenu objects.
-      choroplethMap: React.createRef(),
-      statVarMenuRef: React.createRef(),
-      pc: isPerCapita,
-      // Default to no filtering in stats var sidemenu until API call returns.
-      statsVarFilter: new NoopStatsVarFilter(),
-      // Tracks the currently selected node in sidemenu.
-      statsVarNodes: statsVarNodes,
-    };
-  }
-
-  /**
-   * Toggles the per capita value of the choropleth map and redraws the map.
-   */
-  _togglePerCapita(): void {
-    // Update locally in parent.
-    const newPerCapitaValue = !this.state["pc"];
-    this.setState({
-      pc: newPerCapitaValue,
-    });
-    // Redraw the choropleth map.
-    const choroplethRef = this.state["choroplethMap"].current;
-    choroplethRef.setPerCapita(newPerCapitaValue);
-
-    // Update in URL.
-    const urlParams = new URLSearchParams(window.location.search);
-    urlParams.set("pc", newPerCapitaValue.toString());
-    history.pushState({}, null, "choropleth?" + urlParams.toString());
-  }
-
-  /**
-   * Passes off the downloading and redrawing of a new statistical variable
-   * selection to the child map element.
-   * This function is passed as a callback to the statsvar_menu.
-   * @param statVar
-   */
-  _handleStatVarSelection(statVar: string, statVarLocation: string[]): void {
-    // Update the choropleth map values.
-    const choroplethRef = this.state["choroplethMap"].current;
-    choroplethRef.handleStatVarChange(statVar);
-
-    // Update the displayed value in the stats var sidemenu.
-    const statsVarNodes = {};
-    statsVarNodes[statVar] = { paths: [statVarLocation] };
-    this.setState({ statsVarNodes });
-  }
-
-  // call back function passed down to menu for getting statsVar titles
-  setStatsVarTitle(statsVarId2Title: Record<string, string>): void {
-    this.setState({
-      statsVarTitle: statsVarId2Title,
-    });
-  }
-
-  render(): JSX.Element {
-    return (
-      <div className="choropleth-tool-container">
-        <div className="explore-menu-container" id="explore">
-          <div id="drill-scroll-container">
-            <div className="title">Select variables:</div>
-            <div id="percapita-link" className="text">
-              <label htmlFor="percapita">Per capita</label>
-              <input
-                type="checkbox"
-                id="percapita"
-                checked={this.state["pc"]}
-                name="pc"
-                onClick={this._togglePerCapita.bind(this)}
-              ></input>
-            </div>
-            <Menu
-              ref={this.state && this.state["statVarMenuRef"]}
-              selectedNodes={this.state["statsVarNodes"]}
-              statsVarFilter={this.state["statsVarFilter"]}
-              setStatsVarTitle={this.setStatsVarTitle.bind(this)}
-              addStatsVar={this._handleStatVarSelection.bind(this)}
-              removeStatsVar={() => 0}
-            ></Menu>
-          </div>
-        </div>
-        <div id="main-content">
-          <div id="choropleth-body">
-            <div className="column" id="breadcrumbs"></div>
-            <div id="heading">Loading...</div>
-            <div id="error"></div>
-            <div id="choropleth-section">
-              <div>
-                <ChoroplethMap
-                  ref={this.state && this.state["choroplethMap"]}
-                ></ChoroplethMap>
-              </div>
-              <div id="choropleth-footer">
-                <div id="legend-container">
-                  <div id="date"></div>
-                  <div id="legend"></div>
-                </div>
-              </div>
-              <div id="loading-overlay">
-                <div id="loading-text">Loading...</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-export { MainPane };
